Extract binary observability test helpers
Move run-agent log capture and observability fixtures out of
`main_tests.rs` into a dedicated helper module.

Keep the dispatch tests focused on CLI behaviour and bring the file back
well under the repository line-count limit.
